Add files via upload
Ben,
The commands.png shows how to get this working. Just drop in the test.txt in the same project folder.
The program is pretty much only adding and subtracting.  I was changing it to multiplication and division today, but it'll take me longer than one workday to figure it out.  If you can implement this version, I'll upload the multiply/divide version later and we can just swap out the code.